Hidden non-api objects. Refactoring.  Removed Of suffix for <mutators>. LengthOf.
