+++ src/main/java/org/cactoos/func/IntOf.java
@@ -21,35 +21,49 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.cactoos.text;
+package org.cactoos.func;
 
 import java.io.IOException;
-import org.hamcrest.MatcherAssert;
-import org.hamcrest.Matchers;
-import org.junit.Test;
+import org.cactoos.Scalar;
+import org.cactoos.Text;
+import org.cactoos.text.TextOf;
 
 /**
- * Test case for {@link TextAsFloat}.
+ * Text as {@link Integer}.
+ *
+ * <p>There is no thread-safety guarantee.
  *
  * @author Kirill (g4s8.public@gmail.com)
  * @version $Id$
  * @since 0.2
- * @checkstyle JavadocMethodCheck (500 lines)
  */
-public final class TextAsFloatTest {
+public final class IntOf implements Scalar<Integer> {
 
-    @Test
-    public strictfp void numberTest() throws IOException {
-        MatcherAssert.assertThat(
-            "Can't parse float number",
-            new TextAsFloat("1656.894").value(),
-            // @checkstyle MagicNumber (1 line)
-            Matchers.equalTo(1656.894F)
-        );
+    /**
+     * Source text.
+     */
+    private final Text text;
+
+    /**
+     * Ctor.
+     *
+     * @param string Number-string
+     */
+    public IntOf(final String string) {
+        this(new TextOf(string));
     }
 
-    @Test(expected = NumberFormatException.class)
-    public void failsIfTextDoesNotRepresentAFloat() throws IOException {
-        new TextAsFloat("abc").value();
+    /**
+     * Ctor.
+     *
+     * @param text Number-text
+     */
+    public IntOf(final Text text) {
+        this.text = text;
+    }
+
+    @Override
+    public Integer value() throws IOException {
+        return Integer.valueOf(this.text.asString());
     }
 }
