+++ src/test/java/org/cactoos/func/BoolOfTest.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.cactoos.text;
+package org.cactoos.func;
 
 import java.io.IOException;
 import org.hamcrest.MatcherAssert;
@@ -29,27 +29,39 @@ import org.hamcrest.Matchers;
 import org.junit.Test;
 
 /**
- * Test case for {@link TextAsLong}.
+ * Test case for {@link BoolOf}.
  *
  * @author Kirill (g4s8.public@gmail.com)
  * @version $Id$
  * @since 0.2
  * @checkstyle JavadocMethodCheck (500 lines)
  */
-public final class TextAsLongTest {
+public final class BoolOfTest {
 
     @Test
-    public void numberTest() throws IOException {
+    public void trueTest() throws IOException {
         MatcherAssert.assertThat(
-            "Can't parse long number",
-            new TextAsLong("186789235425346").value(),
-            // @checkstyle MagicNumber (1 line)
-            Matchers.equalTo(186789235425346L)
+            "Can't parse 'true' string",
+            new BoolOf("true").value(),
+            Matchers.equalTo(true)
         );
     }
 
-    @Test(expected = NumberFormatException.class)
-    public void failsIfTextDoesNotRepresentALong() throws IOException {
-        new TextAsLong("abc").value();
+    @Test
+    public void falseTest() throws IOException {
+        MatcherAssert.assertThat(
+            "Can't parse 'false' string",
+            new BoolOf("false").value(),
+            Matchers.equalTo(false)
+        );
+    }
+
+    @Test
+    public void isFalseIfTextDoesNotRepresentABoolean() throws IOException {
+        MatcherAssert.assertThat(
+            "Can't parse a non-boolean string",
+            new BoolOf("abc").value(),
+            Matchers.equalTo(false)
+        );
     }
 }
